Issue KN-10 fix: Moved test cases to SearchBaseActorTest
